primary key unit test updated and hotspot detection added
