+++ webv2/primarykey/primarykey_test.go
@@ -76,18 +76,28 @@ func TestUpdatePrimaryKey(t *testing.T) {
 	handler := http.HandlerFunc(PrimaryKey)
 	handler.ServeHTTP(rr, req)
 
-	var res PrimaryKeyResponse
+	res := &internal.Conv{}
+
 	json.Unmarshal(rr.Body.Bytes(), &res)
 
-	expected := PrimaryKeyResponse{
-		TableId:      1,
-		Columns:      []Column{{ColumnId: 1, ColName: "film_id", Desc: true, Order: 1}},
-		PrimaryKeyId: 1,
+	expectedConv := &internal.Conv{
+
+		SpSchema: map[string]ddl.CreateTable{
+			"film_actor": {
+				Name:     "film_actor",
+				ColNames: []string{"film_id", "actor_id", "last_update"},
+				ColDefs: map[string]ddl.ColumnDef{
+					"film_id":     {Name: "film_id", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 1},
+					"actor_id":    {Name: "actor_id", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 2},
+					"last_update": {Name: "last_update", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 3},
+				},
+				Pks:          []ddl.IndexKey{{Col: "film_id", Order: 1, Desc: true}},
+				Id:           1,
+				PrimaryKeyId: 1,
+			}},
 	}
 
-	assert.Equal(t, expected.TableId, res.TableId)
-	assert.Equal(t, expected.Columns, res.Columns)
-	assert.Equal(t, expected, res)
+	assert.Equal(t, expectedConv, res)
 }
 
 func TestAddPrimaryKey(t *testing.T) {
@@ -136,7 +146,7 @@ func TestAddPrimaryKey(t *testing.T) {
 	handler := http.HandlerFunc(PrimaryKey)
 	handler.ServeHTTP(rr, req)
 
-	var res PrimaryKeyResponse
+	res := &internal.Conv{}
 	json.Unmarshal(rr.Body.Bytes(), &res)
 
 	if status := rr.Code; status != http.StatusOK {
@@ -144,15 +154,24 @@ func TestAddPrimaryKey(t *testing.T) {
 			status, http.StatusOK)
 	}
 
-	expected := PrimaryKeyResponse{
-		TableId:      1,
-		Columns:      []Column{{ColumnId: 1, ColName: "film_id", Desc: true, Order: 1}, {ColumnId: 2, ColName: "actor_id", Desc: false, Order: 2}},
-		PrimaryKeyId: 1,
+	expectedConv := &internal.Conv{
+
+		SpSchema: map[string]ddl.CreateTable{
+			"film_actor": {
+				Name:     "film_actor",
+				ColNames: []string{"film_id", "actor_id", "last_update"},
+				ColDefs: map[string]ddl.ColumnDef{
+					"film_id":     {Name: "film_id", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 1},
+					"actor_id":    {Name: "actor_id", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 2},
+					"last_update": {Name: "last_update", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 3},
+				},
+				Pks:          []ddl.IndexKey{{Col: "film_id", Order: 1, Desc: true}, {Col: "actor_id", Order: 2, Desc: false}},
+				Id:           1,
+				PrimaryKeyId: 1,
+			}},
 	}
 
-	assert.Equal(t, expected.TableId, res.TableId)
-	assert.Equal(t, expected.Columns, res.Columns)
-	assert.Equal(t, expected, res)
+	assert.Equal(t, expectedConv, res)
 }
 
 func TestRemovePrimaryKey(t *testing.T) {
@@ -201,7 +220,7 @@ func TestRemovePrimaryKey(t *testing.T) {
 	handler := http.HandlerFunc(PrimaryKey)
 	handler.ServeHTTP(rr, req)
 
-	var res PrimaryKeyResponse
+	res := &internal.Conv{}
 	json.Unmarshal(rr.Body.Bytes(), &res)
 
 	if status := rr.Code; status != http.StatusOK {
@@ -209,15 +228,23 @@ func TestRemovePrimaryKey(t *testing.T) {
 			status, http.StatusOK)
 	}
 
-	expected := PrimaryKeyResponse{
-		TableId:      1,
-		Columns:      []Column{{ColumnId: 1, ColName: "film_id", Desc: true, Order: 1}, {ColumnId: 2, ColName: "actor_id", Desc: true, Order: 2}},
-		PrimaryKeyId: 1,
-	}
+	expectedConv := &internal.Conv{
 
-	assert.Equal(t, expected.TableId, res.TableId)
-	assert.Equal(t, expected.Columns, res.Columns)
-	assert.Equal(t, expected, res)
+		SpSchema: map[string]ddl.CreateTable{
+			"film_actor": {
+				Name:     "film_actor",
+				ColNames: []string{"film_id", "actor_id", "last_update"},
+				ColDefs: map[string]ddl.ColumnDef{
+					"film_id":     {Name: "film_id", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 1},
+					"actor_id":    {Name: "actor_id", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 2},
+					"last_update": {Name: "last_update", T: ddl.Type{Name: ddl.String, Len: ddl.MaxLength}, Id: 3},
+				},
+				Pks:          []ddl.IndexKey{{Col: "film_id", Order: 1, Desc: true}, {Col: "actor_id", Order: 2, Desc: true}},
+				Id:           1,
+				PrimaryKeyId: 1,
+			}},
+	}
+	assert.Equal(t, expectedConv, res)
 }
 
 func TestPrimarykey(t *testing.T) {
